# src/modules/vouchers/dto/update-voucher.dto.ts
// update-voucher.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreateVoucherDto } from './create-voucher.dto';

export class UpdateVoucherDto extends PartialType(CreateVoucherDto) {}